refactor: replace react-scripts with vite in custom UI
react-scripts (CRA) does not support TypeScript 5. Migrated the
frontend build tooling to Vite with vite-plugin-svgr for SVG
component imports.
