+++ static/onlyoffice-confluence-cloud-custom-ui/index.html
@@ -32,5 +32,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
