+++ static/onlyoffice-confluence-cloud-custom-ui/package.json
@@ -2,28 +2,27 @@
   "name": "@onlyoffice/confluence-cloud-custom-ui",
   "version": "0.1.57",
   "private": true,
-  "homepage": ".",
   "dependencies": {
-    "@atlaskit/app-provider": "^3.1.0",
+    "@atlaskit/app-provider": "^4.1.0",
     "@atlaskit/breadcrumbs": "^15.3.16",
     "@atlaskit/button": "^23.4.1",
     "@atlaskit/css": "^0.19.0",
-    "@atlaskit/css-reset": "^6.6.2",
+    "@atlaskit/css-reset": "^7.3.12",
     "@atlaskit/dropdown-menu": "^16.3.23",
     "@atlaskit/dynamic-table": "^18.2.10",
     "@atlaskit/empty-state": "^10.1.4",
     "@atlaskit/flag": "^17.3.2",
     "@atlaskit/form": "^15.2.0",
     "@atlaskit/heading": "^5.2.16",
-    "@atlaskit/icon": "^28.0.0",
+    "@atlaskit/icon": "^31.0.0",
     "@atlaskit/modal-dialog": "^14.8.0",
     "@atlaskit/popup": "^4.9.0",
-    "@atlaskit/primitives": "^16.4.2",
+    "@atlaskit/primitives": "^18.0.0",
     "@atlaskit/select": "^21.6.3",
     "@atlaskit/tabs": "^18.3.1",
     "@atlaskit/textfield": "^8.2.0",
     "@atlaskit/visually-hidden": "^3.0.4",
-    "@forge/bridge": "5.8.0",
+    "@forge/bridge": "5.13.0",
     "bytes": "^3.1.2",
     "react": "^18",
     "react-dom": "^18"
@@ -32,30 +31,14 @@
     "@types/bytes": "^3.1.5",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "typescript": "4.9.5"
-  },
-  "overrides": {
-    "@typescript-eslint/eslint-plugin": "file:../../node_modules/@typescript-eslint/eslint-plugin",
-    "@typescript-eslint/parser": "file:../../node_modules/@typescript-eslint/parser",
-    "eslint-plugin-react": "file:../../node_modules/eslint-plugin-react",
-    "eslint-plugin-import": "file:../../node_modules/eslint-plugin-import"
+    "@vitejs/plugin-react": "^4.3.4",
+    "typescript": "5.9.3",
+    "vite": "^6.1.0",
+    "vite-plugin-svgr": "^4.3.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   }
 }
